+++ fsq/api.go
@@ -12,6 +12,7 @@ import (
 	"net/url"
 	"strings"
 	"time"
+	"unicode/utf8"
 
 	"golang.org/x/oauth2"
 )
@@ -28,6 +29,7 @@ const (
 )
 
 var errFoursquareResponseTooLarge = errors.New("foursquare response body exceeds 2 MiB")
+var errFoursquareResponseInvalidUTF8 = errors.New("foursquare response body was not valid UTF-8")
 
 func RefuseRedirect(_ *http.Request, _ []*http.Request) error {
 	return http.ErrUseLastResponse
@@ -232,6 +234,9 @@ func decodeFoursquareResponse(body io.Reader, target interface{}) error {
 	}
 	if len(data) > maxFoursquareResponseBytes {
 		return errFoursquareResponseTooLarge
+	}
+	if !utf8.Valid(data) {
+		return errFoursquareResponseInvalidUTF8
 	}
 	response := new(Response)
 	if err := json.Unmarshal(data, response); err != nil {